+++ clients/spotifyApiClient.js
