+++ clients/spotify_client_test.js
@@ -1,9 +1,7 @@
-const SpotifyAPIClient = require("./spotifyApiClient");
+const SpotifyApiClient = require("./spotifyApiClient");
 require("dotenv").config();
 const client_id = process.env.SP_CLIENT_ID;
 const client_secret = process.env.SP_CLIENT_SECRET;
-console.log("clientid", client_id);
-console.log("clientSecret", client_secret);
 
 const client = new SpotifyApiClient(client_id, client_secret);
