+++ package.json
@@ -7,6 +7,8 @@
     "dev": "nodemon index.js"
   },
   "dependencies": {
+    "body-parser": "^1.20.3",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.0",
     "express": "^4.18.2"
   },
